+++ pysys/baserunner.py
@@ -47,14 +47,6 @@ import pysys.utils.allocport
 
 global_lock = threading.Lock() # internal, do not use
 
-N_CPUS = 1 # internal, do not use
-try:
-	# multiprocessing is a new module in 2.6 so we can't assume it
-	import multiprocessing
-	N_CPUS = multiprocessing.cpu_count()
-except ImportError:
-	pass
-
 class BaseRunner(ProcessUser):
 	"""A single instance of the runner class is responsible for orchestrating 
 	concurrent execution of tests, and managing setup and cleanup of 
@@ -166,8 +158,7 @@ class BaseRunner(ProcessUser):
 		
 		self.setKeywordArgs(xargs)
 
-		if self.threads <= 0:
-			self.threads = int(os.getenv('PYSYS_DEFAULT_THREADS', int(float(os.getenv('PYSYS_DEFAULT_THREADS_PER_CPU','1.0'))*N_CPUS)))
+		if len(descriptors)==1: self.threads = 1
 		if self.threads > 1: log.info('Running tests with %d threads', self.threads)
 	
 		self.writers = []